# frontend/src/api/openapi.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "Ground Truth Curator",
		"version": "1.0.6"
	},
	"paths": {
...
		"/v1/assignments/self-serve": {
			"post": {
				"tags": ["assignments"],
				"summary": "Self Serve Assignments",
				"operationId": "self_serve_assignments_v1_assignments_self_serve_post",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"additionalProperties": true,
								"type": "object",
								"title": "Body"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SelfServeResponse"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
...
		"/v1/assignments/{dataset}/{bucket}/{item_id}": {
			"put": {
				"tags": ["assignments"],
				"summary": "Update Item",
				"operationId": "update_item_v1_assignments__dataset___bucket___item_id__put",
				"parameters": [
					{
						"name": "dataset",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Dataset"
						}
					},
					{
						"name": "bucket",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid",
							"title": "Bucket"
						}
					},
					{
						"name": "item_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Item Id"
						}
					},
					{
						"name": "If-Match",
						"in": "header",
						"required": false,
						"schema": {
							"anyOf": [
								{
									"type": "string"
								},
								{
									"type": "null"
								}
							],
							"title": "If-Match"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AssignmentUpdateRequest"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GroundTruthItem-Output"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
		"/v1/assignments/{dataset}/{bucket}/{item_id}/assign": {
			"post": {
				"tags": ["assignments"],
				"summary": "Assign Item",
				"description": "Assign a specific ground truth item to the current user.\n\nThis endpoint:\n- Assigns the item to the current user\n- Sets the item status to draft (even if previously approved/deleted/skipped)\n- Creates an assignment document in the assignments container\n- Returns the updated ground truth item",
				"operationId": "assign_item_v1_assignments__dataset___bucket___item_id__assign_post",
				"parameters": [
					{
						"name": "dataset",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Dataset"
						}
					},
					{
						"name": "bucket",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid",
							"title": "Bucket"
						}
					},
					{
						"name": "item_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Item Id"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GroundTruthItem-Output"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
		"/v1/assignments/{dataset}/{bucket}/{item_id}/duplicate": {
			"post": {
				"tags": ["assignments"],
				"summary": "Duplicate Assignment Item",
				"description": "Duplicate an existing item as a draft rephrase, assign to caller, and return the new item.",
				"operationId": "duplicate_assignment_item_v1_assignments__dataset___bucket___item_id__duplicate_post",
				"parameters": [
					{
						"name": "dataset",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Dataset"
						}
					},
					{
						"name": "bucket",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid",
							"title": "Bucket"
						}
					},
					{
						"name": "item_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Item Id"
						}
					}
				],
				"responses": {
					"201": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GroundTruthItem-Output"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"AssignmentUpdateRequest": {
				"properties": {
					"editedQuestion": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Editedquestion"
					},
					"answer": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Answer"
					},
					"comment": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Comment"
					},
					"status": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/GroundTruthStatus"
							},
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Status"
					},
					"refs": {
						"anyOf": [
							{
								"items": {
									"$ref": "#/components/schemas/Reference"
								},
								"type": "array"
							},
							{
								"type": "null"
							}
						],
						"title": "Refs"
					},
					"manualTags": {
						"anyOf": [
							{
								"items": {
									"type": "string"
								},
								"type": "array"
							},
							{
								"type": "null"
							}
						],
						"title": "Manualtags"
					},
					"approve": {
						"anyOf": [
							{
								"type": "boolean"
							},
							{
								"type": "null"
							}
						],
						"title": "Approve"
					},
					"etag": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Etag"
					},
					"history": {
						"anyOf": [
							{
								"items": {
									"additionalProperties": true,
									"type": "object"
								},
								"type": "array"
							},
							{
								"type": "null"
							}
						],
						"title": "History"
					}
				},
				"additionalProperties": true,
				"type": "object",
				"title": "AssignmentUpdateRequest",
				"description": "Payload for SME update (save draft / approve / skip / delete).\n\nUsing a Pydantic model allows camelCase -> snake_case alias handling. All fields optional; we\nonly mutate those explicitly provided (tracked via model_fields_set)."
			},
...
			"ExpectedBehavior": {
				"type": "string",
				"enum": [
					"tool:search",
					"generation:answer",
					"generation:need-context",
					"generation:clarification",
					"generation:out-of-domain"
				],
				"title": "ExpectedBehavior",
				"description": "Expected behavior tags for history items in ground truth evaluation.\n\nThese tags describe what the agent should do at each turn of a conversation:\n- tool:search: Agent should perform a search/retrieval operation\n- generation:answer: Agent should generate a direct answer\n- generation:need-context: Agent should ask for more context\n- generation:clarification: Agent should ask for clarification\n- generation:out-of-domain: Agent should indicate the query is out of domain"
			},
...
			"GroundTruthItem-Output": {
				"properties": {
					"id": {
						"type": "string",
						"title": "Id"
					},
					"datasetName": {
						"type": "string",
						"title": "Datasetname"
					},
					"bucket": {
						"anyOf": [
							{
								"type": "string",
								"format": "uuid"
							},
							{
								"type": "null"
							}
						],
						"title": "Bucket"
					},
					"status": {
						"$ref": "#/components/schemas/GroundTruthStatus",
						"default": "draft"
					},
					"docType": {
						"type": "string",
						"title": "Doctype",
						"default": "ground-truth-item"
					},
					"schemaVersion": {
						"type": "string",
						"title": "Schemaversion",
						"default": "v2"
					},
					"synthQuestion": {
						"type": "string",
						"title": "Synthquestion"
					},
					"editedQuestion": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Editedquestion"
					},
					"answer": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Answer"
					},
					"refs": {
						"items": {
							"$ref": "#/components/schemas/Reference"
						},
						"type": "array",
						"title": "Refs"
					},
					"manualTags": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"title": "Manualtags"
					},
					"computedTags": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"title": "Computedtags"
					},
					"comment": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Comment"
					},
					"history": {
						"anyOf": [
							{
								"items": {
									"$ref": "#/components/schemas/HistoryItem"
								},
								"type": "array"
							},
							{
								"type": "null"
							}
						],
						"title": "History"
					},
					"contextUsedForGeneration": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Contextusedforgeneration"
					},
					"contextSource": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Contextsource"
					},
					"modelUsedForGeneration": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Modelusedforgeneration"
					},
					"semanticClusterNumber": {
						"anyOf": [
							{
								"type": "integer"
							},
							{
								"type": "null"
							}
						],
						"title": "Semanticclusternumber"
					},
					"weight": {
						"anyOf": [
							{
								"type": "number"
							},
							{
								"type": "null"
							}
						],
						"title": "Weight"
					},
					"samplingBucket": {
						"anyOf": [
							{
								"type": "integer"
							},
							{
								"type": "null"
							}
						],
						"title": "Samplingbucket"
					},
					"questionLength": {
						"anyOf": [
							{
								"type": "integer"
							},
							{
								"type": "null"
							}
						],
						"title": "Questionlength"
					},
					"assignedTo": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Assignedto"
					},
					"assignedAt": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "null"
							}
						],
						"title": "Assignedat"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time",
						"title": "Updatedat"
					},
					"updatedBy": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Updatedby"
					},
					"reviewedAt": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "null"
							}
						],
						"title": "Reviewedat"
					},
					"_etag": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Etag"
					},
					"totalReferences": {
						"type": "integer",
						"title": "Totalreferences",
						"default": 0
					},
					"tags": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"title": "Tags",
						"description": "Merged view of manualTags and computedTags for API responses.\n\nReturns the union of manual and computed tags, sorted for consistency.",
						"readOnly": true
					}
				},
				"type": "object",
				"required": ["id", "datasetName", "synthQuestion", "tags"],
				"title": "GroundTruthItem",
				"description": "Canonical Ground Truth item aligned to wire schema (schemaVersion v1).\n\nAll fields with camelCase wire names use aliases; we accept both field names and aliases\non input (populate_by_name=True) and always serialize using by_alias."
			},
...
			"GroundTruthStatus": {
				"type": "string",
				"enum": ["draft", "approved", "deleted", "skipped"],
				"title": "GroundTruthStatus"
			},
			"HTTPValidationError": {
				"properties": {
					"detail": {
						"items": {
							"$ref": "#/components/schemas/ValidationError"
						},
						"type": "array",
						"title": "Detail"
					}
				},
				"type": "object",
				"title": "HTTPValidationError"
			},
			"HistoryItem": {
				"properties": {
					"role": {
						"$ref": "#/components/schemas/HistoryItemRole"
					},
					"msg": {
						"type": "string",
						"title": "Msg"
					},
					"refs": {
						"anyOf": [
							{
								"items": {
									"$ref": "#/components/schemas/Reference"
								},
								"type": "array"
							},
							{
								"type": "null"
							}
						],
						"title": "Refs"
					},
					"expectedBehavior": {
						"anyOf": [
							{
								"items": {
									"$ref": "#/components/schemas/ExpectedBehavior"
								},
								"type": "array"
							},
							{
								"type": "null"
							}
						],
						"title": "Expectedbehavior",
						"description": "Expected behavior(s) for this turn in the conversation (e.g., tool:search, generation:answer)"
					}
				},
				"type": "object",
				"required": ["role", "msg"],
				"title": "HistoryItem",
				"description": "Represents a single item in the multi-turn history."
			},
			"HistoryItemRole": {
				"type": "string",
				"enum": ["user", "assistant"],
				"title": "HistoryItemRole"
			},
...
			"Reference": {
				"properties": {
					"url": {
						"type": "string",
						"title": "Url",
						"description": "Reference URL (required, non-empty)"
					},
					"title": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Title",
						"description": "Human-readable title for the reference"
					},
					"content": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Content"
					},
					"keyExcerpt": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Keyexcerpt"
					},
					"type": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Type"
					},
					"bonus": {
						"type": "boolean",
						"title": "Bonus",
						"default": false
					},
					"messageIndex": {
						"anyOf": [
							{
								"type": "integer"
							},
							{
								"type": "null"
							}
						],
						"title": "Messageindex"
					}
				},
				"type": "object",
				"required": ["url"],
				"title": "Reference",
				"description": "Wire reference object.\n\n{ url, title, content, keyExcerpt, type, bonus, messageIndex }"
			},
...
			"SelfServeResponse": {
				"properties": {
					"assigned": {
						"items": {
							"$ref": "#/components/schemas/GroundTruthItem-Output"
						},
						"type": "array",
						"title": "Assigned"
					},
					"requested": {
						"type": "integer",
						"title": "Requested"
					},
					"assignedCount": {
						"type": "integer",
						"title": "Assignedcount"
					}
				},
				"type": "object",
				"required": ["assigned", "requested", "assignedCount"],
				"title": "SelfServeResponse"
			},
...
			"ValidationError": {
				"properties": {
					"loc": {
						"items": {
							"anyOf": [
								{
									"type": "string"
								},
								{
									"type": "integer"
								}
							]
						},
						"type": "array",
						"title": "Location"
					},
					"msg": {
						"type": "string",
						"title": "Message"
					},
					"type": {
						"type": "string",
						"title": "Error Type"
					}
				},
				"type": "object",
				"required": ["loc", "msg", "type"],
				"title": "ValidationError"
			}
		}
	}
}